Added GuideViewModelProtocol
